Remove page object, change template render function.
Update the pybald version number to 0.2.0 in prep for tagging. Change
the template rendering function to take an explicit template name rather
than depend on a key in the data dictionary. Cleanup the BaseController
to make it even more optional. Fix the error controller to use the new
template rendering call sig. Add the finally call to session.remove in
db_middleware. This should be good enough to always close the session
regardless of any other exceptions that occur. Add the 'page' helper
and include in all template rendering. No longer need a page object
in the controller to handle HTML helper functions for static assets.