Create new models classes that are bound to the pybald context
Now every time pybald.configure is called, the context creates new
context-specific classes to bind to models. This simplifies sharing
of resources (there is no sharing) and less keeping track of what
each of the contexts models have loaded. When a context is active
context.models.Model will be bound to a class that creates
instances bound to that configuration with their own scoped session
and database engine.

This should, in theory, have some performance implications since the
attributes that were originally the 'switch points' for contexts are
now all fixed removing the need to switch at runtime what engine
or metadata is being used.