Include object with general exception dispatches?
We noticed that PyBald passes the code and the raw
exception when dispatching HTTPExceptions to the
error controller, but only a text representation for other
exceptions...

This seems kinda inconsistent, and is also problematic for
libraries like Sentry that can generate detailed error reports
from an Exception object (but not from a string repr).

I'm guessing there was something behind the current design,
though, probably having to do with avoiding references (and
memory leaks) around the exception objects... This patch is
more to start the Pull Request discussion.  ;-)