Eliminate Smarterer-era "can_login" check
At some point we added the unconditional check for `user.can_login`, and also added the more generic, optional hook for  `user.valid`.

Having gone over this with the various teams using PyBald, it seems like we should have ditched `user.can_login` ages ago, but just kind of forgot about it...

I believe only Smarterer relies on the `can_login` behavior currently, and it's been pinned to an ancient version of PyBald (and likely will continue to be that way), so it seems safe to get rid of in Master.

My only question is, in addition to setting `environ['REMOTE_USER']` to None when `valid` exists and is Falsey, should we also set session.user to None? Or is that a step too far?